Refactor database migration and seeding process
- Updated docker-compose.yml to change database connection string and SMTP server.
- Added a new service for database migration in docker-compose.
- Introduced a new migration script (src/db/migrate.ts) to handle database migrations and seeding.
- Removed the old seeder script (src/db/seeder.ts) as its functionality is now integrated into the migration script.
- Updated package.json to include new npm scripts for database migration.
- Upgraded tsx package to version 4.22.3 in package-lock.json and package.json.
